move function to partition file
